move: notification test
