+++ pkg/server/alert/alert_analyze_test.go
@@ -16,7 +16,6 @@ import (
 	"github.com/ca-risken/core/proto/project"
 	projectmock "github.com/ca-risken/core/proto/project/mocks"
 	"github.com/golang/protobuf/ptypes/empty"
-	"github.com/jarcoal/httpmock"
 	"gorm.io/gorm"
 )
 
@@ -103,121 +102,6 @@ func TestAnalyzeAlert(t *testing.T) {
 			}
 			if !reflect.DeepEqual(got, c.want) {
 				t.Fatalf("Unexpected response: want=%+v, got=%+v", c.want, got)
-			}
-		})
-	}
-}
-
-func TestNotificationAlert(t *testing.T) {
-	httpmock.Activate()
-	defer httpmock.DeactivateAndReset()
-	now := time.Now()
-	mockDB := mocks.MockAlertRepository{}
-	svc := AlertService{repository: &mockDB}
-
-	httpmock.RegisterResponder("POST", "http://hogehoge.com", httpmock.NewStringResponder(200, "mocked"))
-	httpmock.RegisterResponder("POST", "http://fugafuga.com", httpmock.NewErrorResponder(errors.New("Something Wrong")))
-	cases := []struct {
-		name                               string
-		alertCondition                     *model.AlertCondition
-		alert                              *model.Alert
-		wantErr                            bool
-		mockListAlertCondNotification      *[]model.AlertCondNotification
-		mockListAlertCondNotificationErr   error
-		mockGetNotification                *model.Notification
-		mockGetNotificationErr             error
-		mockGetProject                     *project.Project
-		mockGetProjectErr                  error
-		mockUpsertAlertCondNotification    *model.AlertCondNotification
-		mockUpsertAlertCondNotificationErr error
-	}{
-		{
-			name:                             "OK 0 AlertCondNotification",
-			alertCondition:                   &model.AlertCondition{AlertConditionID: 1},
-			alert:                            &model.Alert{},
-			wantErr:                          false,
-			mockListAlertCondNotification:    &[]model.AlertCondNotification{},
-			mockListAlertCondNotificationErr: nil,
-		},
-		{
-			name:                               "OK Notification Success",
-			alertCondition:                     &model.AlertCondition{AlertConditionID: 1},
-			alert:                              &model.Alert{},
-			wantErr:                            false,
-			mockListAlertCondNotification:      &[]model.AlertCondNotification{{AlertConditionID: 1, NotificationID: 1}},
-			mockListAlertCondNotificationErr:   nil,
-			mockGetNotification:                &model.Notification{Type: "slack", NotifySetting: `{"webhook_url":"http://hogehoge.com"}`},
-			mockGetNotificationErr:             nil,
-			mockGetProject:                     &project.Project{},
-			mockGetProjectErr:                  nil,
-			mockUpsertAlertCondNotification:    &model.AlertCondNotification{},
-			mockUpsertAlertCondNotificationErr: nil,
-		},
-		{
-			name:                             "OK Don't send Notification caused NotifedAt",
-			alertCondition:                   &model.AlertCondition{AlertConditionID: 1},
-			alert:                            &model.Alert{},
-			wantErr:                          false,
-			mockListAlertCondNotification:    &[]model.AlertCondNotification{{AlertConditionID: 1, NotificationID: 1, CacheSecond: 30, NotifiedAt: now}},
-			mockListAlertCondNotificationErr: nil,
-			mockGetNotification:              &model.Notification{Type: "slack", NotifySetting: `{"webhook_url":"http://fugafuga.com"}`},
-			mockGetNotificationErr:           nil,
-		},
-		{
-			name:                             "Error ListAlertCondNotification Failed",
-			alertCondition:                   &model.AlertCondition{AlertConditionID: 1},
-			alert:                            &model.Alert{},
-			wantErr:                          true,
-			mockListAlertCondNotification:    nil,
-			mockListAlertCondNotificationErr: errors.New("Somethinng error occured"),
-		},
-		{
-			name:                             "Error GetNotification Failed",
-			alertCondition:                   &model.AlertCondition{AlertConditionID: 1},
-			alert:                            &model.Alert{},
-			wantErr:                          true,
-			mockListAlertCondNotification:    &[]model.AlertCondNotification{{AlertConditionID: 1, NotificationID: 1}},
-			mockListAlertCondNotificationErr: nil,
-			mockGetNotification:              nil,
-			mockGetNotificationErr:           errors.New("Somethinng error occured"),
-		},
-		{
-			name:                             "Error GetNotification Failed",
-			alertCondition:                   &model.AlertCondition{AlertConditionID: 1},
-			alert:                            &model.Alert{},
-			wantErr:                          true,
-			mockListAlertCondNotification:    &[]model.AlertCondNotification{{AlertConditionID: 1, NotificationID: 1}},
-			mockListAlertCondNotificationErr: nil,
-			mockGetNotification:              nil,
-			mockGetNotificationErr:           errors.New("Somethinng error occured"),
-			mockGetProject:                   nil,
-			mockGetProjectErr:                errors.New("Somethinng error occured"),
-		},
-		{
-			name:                               "Error UpsertAlertCondNotification Failed",
-			alertCondition:                     &model.AlertCondition{AlertConditionID: 1},
-			alert:                              &model.Alert{},
-			wantErr:                            true,
-			mockListAlertCondNotification:      &[]model.AlertCondNotification{{AlertConditionID: 1, NotificationID: 1}},
-			mockListAlertCondNotificationErr:   nil,
-			mockGetNotification:                &model.Notification{Type: "slack", NotifySetting: `{"webhook_url":"http://hogehoge.com"}`},
-			mockGetNotificationErr:             nil,
-			mockGetProject:                     &project.Project{},
-			mockGetProjectErr:                  nil,
-			mockUpsertAlertCondNotification:    nil,
-			mockUpsertAlertCondNotificationErr: errors.New("Somethinng error occured"),
-		},
-	}
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			mockDB = mocks.MockAlertRepository{}
-			mockDB.On("ListAlertCondNotification").Return(c.mockListAlertCondNotification, c.mockListAlertCondNotificationErr).Once()
-			mockDB.On("GetNotification").Return(c.mockGetNotification, c.mockGetNotificationErr).Once()
-			mockDB.On("UpsertAlertCondNotification").Return(c.mockUpsertAlertCondNotification, c.mockUpsertAlertCondNotificationErr).Once()
-			mockDB.On("GetProject").Return(c.mockGetProject, c.mockGetProjectErr).Once()
-			got := svc.NotificationAlert(context.Background(), c.alertCondition, c.alert, &[]model.AlertRule{}, &project.Project{})
-			if (got != nil && !c.wantErr) || (got == nil && c.wantErr) {
-				t.Fatalf("Unexpected error: %+v", got)
 			}
 		})
 	}
